Add relevant repositories to the federation.
WARNING: This state of the federation is not yet ready for use in production since a) no project uses the federation yet and b) the versions of most of the repositories haven't been tested together.

However, the presence of these function is required in order to start onboarding member projects.

Once this commit has been merged, we can start updating member projects to use the federation via the the required internal_deps.bzl and internal_setup.bzl files.
Next, we'll update the federation to use the latest versions of these projects, and we'll also fix their setup functions.